changed double to long
upping that commit number :)

+++ app/src/main/java/com/a0xffffffff/dinesum/User.java
@@ -22,7 +22,7 @@ public class User {
         mPoints = 50;                           //set rating to 0?
     }
 
-    public double getPoints()
+    public long getPoints()
     {
         return mPoints;
     }
